requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Babel==2.12.1
cfgv==3.3.1
charset-normalizer==3.1.0
discord-py-interactions==5.0.1
discord-py-slash-command==4.2.1
discord-typings==0.5.1
discord.py==2.2.2
distlib==0.3.6
emoji==2.2.0
filelock==3.12.0
frozenlist==1.3.3
greenlet==2.0.2
identify==2.5.22
idna==3.4
multidict==6.0.4
nest-asyncio==1.5.6
nodeenv==1.7.0
platformdirs==3.2.0
pre-commit==3.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.21.0
yarl==1.8.2
